[FIO internal] caam: imx7ulp: RPMB workaround for closed boards
Due to a problem in one of our releases, some imx7ulp boards were able
to access RPMB while still in the open state.

In the open state, CAAM provides a HUK that is different to the one
after the board has been closed.

This means that once the boards were closed, RPMB would not longer be
accessible.

The current commit addresses that situation.