[FIO toup] core: increase stacks when using SE050
The se050 session is initiated by the op-tee core (not a TA) and this
already requires larger stacks to support the allocation of the i2c
messages done by the library.

It is very possible that the stack increases in this commit are on the
larger size but it saves us the time of having to go through the TA's
tunning them.

Signed-off-by: Jorge Ramirez-Ortiz <jorge@foundries.io>